prodDowntimes: changed the corroboration process.
User with the `PROD_DATA:MANAGE` privilege can always change status, reason and AOR.
User without the `PROD_DOWNTIMES:MANAGE` privilege can only comment.
If the downtime is undecided, user with access to the current AOR can reject or confirm the downtime. When rejected, the downtime's AOR is changed to the default AOR of the subdivision the downtime belongs to.
If the downtime is rejected, user with access to the current AOR can assign a new reason and AOR while changing the status to undecided.
If the downtime is confirmed, the status cannot be changed.